Move IL patch methods and attribute declarations to respective classes where possible
