+++ Celeste.Mod.mm/Patches/DeathEffect.cs
@@ -1,5 +1,9 @@
+using System;
 using Microsoft.Xna.Framework;
+using Mono.Cecil;
+using Mono.Cecil.Cil;
 using MonoMod;
+using MonoMod.Cil;
 
 namespace Celeste {
     class patch_DeathEffect : DeathEffect {
@@ -16,5 +20,23 @@ namespace Celeste {
             if (Entity != null)
                 Draw(Entity.Position + Position, Color, Percent);
         }
+    }
+}
+
+namespace MonoMod {
+    /// <summary>
+    /// Patch DeathEffect.Update to fix death effects never get removed
+    /// </summary>
+    [MonoModCustomMethodAttribute(nameof(MonoModRules.PatchDeathEffectUpdate))]
+    class PatchDeathEffectUpdateAttribute : Attribute { }
+
+    static partial class MonoModRules {
+
+        public static void PatchDeathEffectUpdate(ILContext context, CustomAttribute attrib) {
+            ILCursor cursor = new ILCursor(context);
+            cursor.GotoNext(instr => instr.OpCode == OpCodes.Ble_Un_S);
+            cursor.Next.OpCode = OpCodes.Blt_Un_S;
+        }
+
     }
 }
