+++ Celeste.Mod.mm/Patches/GoldenBlock.cs
@@ -1,8 +1,14 @@
 ﻿#pragma warning disable CS0626 // Method, operator, or accessor is marked external and has no attributes on it
 
+using System;
 using Microsoft.Xna.Framework;
+using Mono.Cecil;
+using Mono.Cecil.Cil;
 using Monocle;
 using MonoMod;
+using MonoMod.Cil;
+using MonoMod.InlineRT;
+using MonoMod.Utils;
 
 namespace Celeste {
     class patch_GoldenBlock : GoldenBlock {
@@ -27,5 +33,32 @@ namespace Celeste {
         [MonoModIgnore] // we don't want to change anything in the method...
         [PatchGoldenBlockStaticMovers] // ... except manipulating it manually with MonoModRules
         public extern override void Awake(Scene scene);
+    }
+}
+
+namespace MonoMod {
+    /// <summary>
+    /// Patches GoldenBlocks to disable static movers if the block is disabled.
+    /// </summary>
+    [MonoModCustomMethodAttribute(nameof(MonoModRules.PatchGoldenBlockStaticMovers))]
+    class PatchGoldenBlockStaticMoversAttribute : Attribute { }
+
+    static partial class MonoModRules {
+
+        public static void PatchGoldenBlockStaticMovers(ILContext context, CustomAttribute attrib) {
+            TypeDefinition t_Platform = MonoModRule.Modder.Module.GetType("Celeste.Platform");
+            MethodDefinition m_Platform_DisableStaticMovers = t_Platform.FindMethod("System.Void DisableStaticMovers()");
+            MethodDefinition m_Platform_DestroyStaticMovers = t_Platform.FindMethod("System.Void DestroyStaticMovers()");
+
+            ILCursor cursor = new ILCursor(context);
+            cursor.GotoNext(MoveType.After, instr => instr.MatchCall("Celeste.Solid", "Awake"));
+            cursor.Emit(OpCodes.Ldarg_0);
+            cursor.Emit(OpCodes.Call, m_Platform_DisableStaticMovers);
+
+            cursor.GotoNext(instr => instr.MatchCall("Monocle.Entity", "RemoveSelf"));
+            cursor.Emit(OpCodes.Ldarg_0);
+            cursor.Emit(OpCodes.Call, m_Platform_DestroyStaticMovers);
+        }
+
     }
 }
