+++ Celeste.Mod.mm/Patches/Monocle/ErrorLog.cs
@@ -4,6 +4,10 @@ using Celeste.Mod;
 using MonoMod;
 using MonoMod.Utils;
 using System;
+using Mono.Cecil;
+using Mono.Cecil.Cil;
+using MonoMod.Cil;
+using MonoMod.InlineRT;
 
 namespace Monocle {
     class patch_ErrorLog {
@@ -18,6 +22,40 @@ namespace Monocle {
         [MonoModIgnore] // We don't want to change anything about the method...
         [PatchErrorLogWrite] // ... except for manually manipulating the method via MonoModRules
         public static extern void Write(string str);
+
+    }
+}
+
+namespace MonoMod {
+    /// <summary>
+    /// Find ldfld Engine::Version + ToString. Pop ToString result, call Everest::get_VersionCelesteString
+    /// </summary>
+    [MonoModCustomMethodAttribute(nameof(MonoModRules.PatchErrorLogWrite))]
+    class PatchErrorLogWriteAttribute : Attribute { }
+
+    static partial class MonoModRules {
+
+        public static void PatchErrorLogWrite(ILContext context, CustomAttribute attrib) {
+            TypeDefinition t_Everest = MonoModRule.Modder.FindType("Celeste.Mod.Everest").Resolve();
+            MethodDefinition m_Everest_get_VersionCelesteString = t_Everest.FindMethod("System.String get_VersionCelesteString()");
+
+            /* We expect something similar enough to the following:
+            call     class Monocle.Engine Monocle.Engine::get_Instance() // We're here
+            ldfld    class [mscorlib] System.Version Monocle.Engine::Version 
+            callvirt instance string[mscorlib] System.Object::ToString() 
+
+            Note that MonoMod requires the full type names (System.String instead of string)
+            */
+
+            ILCursor cursor = new ILCursor(context);
+            cursor.GotoNext(instr => instr.MatchCall("Monocle.Engine", "get_Instance"),
+                instr => instr.MatchLdfld("Monocle.Engine", "Version"),
+                instr => instr.MatchCallvirt("System.Object", "ToString"));
+
+            // Remove all that and replace with our own string.
+            cursor.RemoveRange(3);
+            cursor.Emit(OpCodes.Call, m_Everest_get_VersionCelesteString);
+        }
 
     }
 }
