+++ Celeste.Mod.mm/Patches/Settings.cs
@@ -1,8 +1,13 @@
-﻿using Microsoft.Xna.Framework.Input;
+﻿using System;
+using Microsoft.Xna.Framework.Input;
 using MonoMod;
 using System.Collections.Generic;
 using System.Linq;
 using System.Xml.Serialization;
+using Mono.Cecil;
+using Mono.Cecil.Cil;
+using MonoMod.Cil;
+using MonoMod.Utils;
 
 namespace Celeste {
     class patch_Settings : Settings {
@@ -159,6 +164,42 @@ namespace Celeste {
         }
 
         #endregion
+
+    }
+}
+
+namespace MonoMod {
+    /// <summary>
+    /// Patches Settings.SetDefaultKeyboardControls so that TranslateKeys only gets called when reset = true.
+    /// </summary>
+    [MonoModCustomMethodAttribute(nameof(MonoModRules.PatchSettingsDoNotTranslateKeys))]
+    class PatchSettingsDoNotTranslateKeysAttribute : Attribute { }
+
+    static partial class MonoModRules {
+
+        public static void PatchSettingsDoNotTranslateKeys(ILContext il, CustomAttribute attrib) {
+            ILCursor c = new ILCursor(il);
+
+            // find the instruction after the group of TranslateKeys.
+            while (c.TryGotoNext(MoveType.After, instr => instr.MatchCall("Celeste.Settings", "TranslateKeys"))) { }
+            Instruction jumpTarget = c.Next;
+
+            c.Index = 0;
+
+            // go just before the first TranslateKeys call.
+            if (c.TryGotoNext(MoveType.AfterLabel,
+                instr => instr.MatchLdarg(0),
+                instr => instr.OpCode == OpCodes.Ldfld,
+                instr => instr.OpCode == OpCodes.Ldfld,
+                instr => instr.MatchCall("Celeste.Settings", "TranslateKeys"))) {
+                // enclose all TranslateKeys inside a if (reset || !Existed).
+                c.Emit(OpCodes.Ldarg_1);
+                c.Emit(OpCodes.Brtrue, c.Next);
+                c.Emit(OpCodes.Ldarg_0);
+                c.Emit(OpCodes.Ldfld, il.Method.DeclaringType.FindField("Existed"));
+                c.Emit(OpCodes.Brtrue, jumpTarget);
+            }
+        }
 
     }
 }
