+++ Celeste.Mod.mm/Patches/TextMenu.cs
@@ -8,6 +8,10 @@ using MonoMod;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using Mono.Cecil;
+using Mono.Cecil.Cil;
+using MonoMod.Cil;
+using MonoMod.Utils;
 
 namespace Celeste {
     public class patch_TextMenu : TextMenu {
@@ -449,6 +453,28 @@ namespace Celeste {
         /// <inheritdoc cref="patch_TextMenu.Remove(TextMenu.Item)"/>
         public static TextMenu Remove(this TextMenu self, TextMenu.Item item)
             => ((patch_TextMenu) self).Remove(item);
+
+    }
+}
+
+namespace MonoMod {
+    /// <summary>
+    /// Patches the unselected color in TextMenu.Option to make it customizable.
+    /// </summary>
+    [MonoModCustomMethodAttribute(nameof(MonoModRules.PatchTextMenuOptionColor))]
+    class PatchTextMenuOptionColorAttribute : Attribute { }
+
+    static partial class MonoModRules {
+
+        public static void PatchTextMenuOptionColor(ILContext context, CustomAttribute attrib) {
+            FieldReference f_UnselectedColor = context.Method.DeclaringType.FindField("UnselectedColor");
+
+            ILCursor cursor = new ILCursor(context);
+            cursor.GotoNext(instr => instr.MatchCall("Microsoft.Xna.Framework.Color", "get_White"));
+            cursor.Emit(OpCodes.Ldarg_0);
+            cursor.Next.OpCode = OpCodes.Ldfld;
+            cursor.Next.Operand = f_UnselectedColor;
+        }
 
     }
 }
